sửa

+++ admin/reply.php
@@ -34,7 +34,7 @@ include("./header.php");
 						</div>
 					</td>
 					<td><?php echo $row['book_descr']; ?></td>
-					<td><a href="delete_books_add.php?bookadid=<?php echo $row['book_ad_id']; ?>" class="text-danger"><i class="fas fa-minus-square"></i></a></td>
+					<td><a href="delete_books_add.php?bookadid=<?php echo $row['book_ad_id']; ?>" onclick="return confirm('Bạn chắc chắn muốn xóa?')" class="text-danger"><i class="fas fa-minus-square"></i></a></td>
 				</tr>
 		<?php
 				$i++;
